backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-firestore==2.14.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pdfplumber==0.10.3
openai==1.10.0
anthropic==0.18.1
instructor==0.5.2
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pydantic-ai
nest_asyncio
